pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dream-job-copilot"
version = "0.1.0"
description = "An agentic AI pipeline to find and apply for your dream job"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["ai", "job-search", "langchain", "langgraph", "resume", "career"]
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "openai>=1.50.0",
    "tavily-python>=0.5.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "rich>=13.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.7.0",
]

[project.scripts]
dream-job-copilot = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
